[FIX] Do away for now with farmer_id number
If really needed, this could be computed from the partner id numbers. Also
is it a privacy sensitive number. Why display that on contracts?
As for the filter on duplicate id numbers, there should really be a check
preventing the addition of these at all. This should be en enhancement to
the partner_identification module.